# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api sistema de agendacion de citas asopormen",
        "description": "Api para la gestion de citas a usuarios de la ips Asopormen.sp",
        "termsOfService": "https://asopormen.org.co/",
        "contact": {
            "name": "Equipo de soporte",
            "email": "msolegario.cisneros@gmail.com"
        },
        "license": {
            "name": "Licencia",
            "url": "https://example.com/license"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://por_definir/api/documentation"
        }
    ],
    "paths": {
...
        "/citas/get_num_citas/{authorization}/{procedim}": {
            "get": {
                "tags": [
                    "Citas"
                ],
                "summary": "Obtener número de citas por autorización y procedimiento que no han sido canceladas",
                "description": "Obtiene el número de citas registradas para una autorización específica y un procedimiento determinado.",
                "operationId": "147b4a3b51bcab23867de23dc3b9c892",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "path",
                        "description": "Número de autorización",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "AUT123456"
                        }
                    },
                    {
                        "name": "procedim",
                        "in": "path",
                        "description": "Código o nombre del procedimiento",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Procedimiento A"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Número de citas encontrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "description": "Número de citas encontradas",
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron citas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "No se encontraron citas"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Server error"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}